backend/requirements.txt
# Core server
fastapi
uvicorn
psutil
aiohttp>=3.9
python-multipart>=0.0.9

# GPU metrics (optional at runtime)
pynvml>=11.5.0

# Vector store + embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document parsing (NO pdfplumber for now)
pymupdf==1.24.9
pdfminer.six==20240706
python-docx>=0.8.11
pandas>=2.2.2
openpyxl>=3.1.5
pypdf

# Make sure any "Crypto" imports resolve to the maintained lib
pycryptodome
